whatwewatching/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mypy_extensions==1.1.0
narwhals==1.46.0
numpy==2.2.6
openai==1.95.0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
